refactor(car VM): move fetch all to exchange selection VM
Remove local state usage, this is a impl error, needs to be stateless between screens
